+++ src/components/molecules/WalkThrough/types.ts
@@ -1,0 +1,14 @@
+export type WalkThroughStep = 'template' | 'resource' | 'syntax' | 'cluster' | 'kustomizeHelm';
+export type WalkThroughContentProps = {
+  data: {
+    step: WalkThroughStep;
+    title: string;
+    content: string;
+  };
+  currentStep: number;
+};
+
+export enum StepEnum {
+  Previous = -1,
+  Next = 1,
+}
